Move references out
