+++ models/experimental/gemma3_4b/tests/vision_tests/test_patch_embedding.py
@@ -15,6 +15,7 @@ import torch
 from models.tt_transformers.tt.model_config import ModelArgs
 
 from models.experimental.gemma3_4b.tt.gemma_conv2d_patch import TtGemmaConv2dPatch
+from models.experimental.gemma3_4b.tests.references import reference_siglip_patch_embed
 from models.utility_functions import comp_allclose, comp_pcc, skip_for_grayskull
 from ttnn import ConcatMeshToTensor
 
@@ -67,7 +68,7 @@ def test_conv2d_inference(
     logger.info(f"Input tensor shape: {input_tensor.shape}")
 
     ##### Perform the torch ops #####
-    reference_model = model_args.reference_siglip_patch_embed()
+    reference_model = reference_siglip_patch_embed(model_args)
     # reference_model.load_state_dict(partial_state_dict)
     reference_output = reference_model(input_tensor)
     del reference_model
