+++ models/experimental/gemma3_4b/tests/vision_tests/test_vision_attention.py
@@ -20,6 +20,7 @@ from models.tt_transformers.tt.model_config import ModelArgs
 
 
 from models.experimental.gemma3_4b.tt.gemma_image_attention import TtGemmaImageAttention
+from models.experimental.gemma3_4b.tests.references import reference_vision_attention
 from models.utility_functions import comp_allclose, comp_pcc, skip_for_grayskull
 
 
@@ -53,7 +54,7 @@ def test_attention_inference(batch, num_chunks, mesh_device, reset_seeds):
 
     dim = model_args.vision_dim
 
-    reference_model = model_args.reference_vision_attention()
+    reference_model = reference_vision_attention(model_args)
     # reference_model.load_state_dict(partial_state_dict)
     reference_model.eval()
