+++ models/experimental/gemma3_4b/tests/vision_tests/test_vision_embedding.py
@@ -15,6 +15,7 @@ import ttnn
 from models.tt_transformers.tt.model_config import ModelArgs
 
 from models.experimental.gemma3_4b.tt.siglip_vision_embedding import TtSiglipVisionEmbeddings
+from models.experimental.gemma3_4b.tests.references import reference_vision_embedding
 from models.utility_functions import comp_allclose, comp_pcc, skip_for_grayskull
 from ttnn import ConcatMeshToTensor
 
@@ -53,7 +54,7 @@ def test_vision_embedding_integration(
 
     input_tensor = torch.randn((bsz, in_channels, image_size, image_size))
 
-    reference_model = model_args.reference_vision_embedding()
+    reference_model = reference_vision_embedding(model_args)
     # reference_model.load_state_dict(partial_state_dict)
     reference_output = reference_model(input_tensor)
