+++ models/experimental/gemma3_4b/tests/vision_tests/test_vision_layernorm.py
@@ -14,6 +14,7 @@ from loguru import logger
 import ttnn
 from models.tt_transformers.tt.model_config import ModelArgs
 from models.tt_transformers.tt.multimodal.llama_layernorm import TtLayerNorm  # Updated import for LayerNorm
+from models.experimental.gemma3_4b.tests.references import reference_vision_layernorm
 from models.utility_functions import comp_allclose, comp_pcc, nearest_32, skip_for_grayskull
 
 
@@ -47,7 +48,7 @@ def test_layernorm_inference(mesh_device, reset_seeds, layer_name):
 
     model_args.WEIGHTS_DTYPE = dtype
     # Reference HF MLP (from Gemma3 vision tower)
-    reference_model = model_args.reference_vision_layernorm(layer_name)
+    reference_model = reference_vision_layernorm(model_args, layer_name)
     # reference_model.load_state_dict(partial_state_dict)
     reference_model.eval()
