+++ models/experimental/gemma3_4b/tests/vision_tests/test_vision_mlp.py
@@ -15,6 +15,7 @@ import ttnn
 from models.tt_transformers.tt.ccl import TT_CCL
 from models.tt_transformers.tt.model_config import ModelArgs
 from models.experimental.gemma3_4b.tt.gemma_image_mlp import TtGemmaImageFeedForward
+from models.experimental.gemma3_4b.tests.references import reference_vision_mlp
 from models.utility_functions import comp_allclose, comp_pcc, nearest_32, skip_for_grayskull
 
 
@@ -49,7 +50,7 @@ def test_mlp_inference(batch, num_chunks, mesh_device, reset_seeds):
 
     dim = model_args.vision_dim
     seq_len = nearest_32(model_args.image_size) * num_chunks
-    reference_model = model_args.reference_vision_mlp()
+    reference_model = reference_vision_mlp(model_args)
     # reference_model.load_state_dict(partial_state_dict)
 
     tt_ccl = TT_CCL(mesh_device)
