+++ models/experimental/gemma3_4b/tests/vision_tests/test_vision_rmsnorm.py
@@ -20,6 +20,8 @@ from models.tt_transformers.tt.ccl import TT_CCL
 from models.utility_functions import comp_allclose, comp_pcc, skip_for_grayskull
 from models.tt_transformers.tt.model_config import ModelArgs
 
+from models.experimental.gemma3_4b.tests.references import reference_vision_rms_norm
+
 
 @torch.no_grad()
 @skip_for_grayskull("Requires wormhole_b0 to run")
@@ -53,7 +55,7 @@ def test_rmsnorm_inference(seq_len, batch_size, reset_seeds, device):
     tt_model_args.n_layers = 1
     state_dict = tt_model_args.load_state_dict()
 
-    reference_model = tt_model_args.reference_vision_rms_norm()  # Gemma3 RMSNorm
+    reference_model = reference_vision_rms_norm(tt_model_args)  # Gemma3 RMSNorm
     first_layer_prefix = "multi_modal_projector.mm_soft_emb_norm."
 
     partial_state_dict = {
