+++ models/experimental/gemma3_4b/tests/vision_tests/test_vision_transformer.py
@@ -14,6 +14,7 @@ import ttnn
 from models.tt_transformers.tt.ccl import TT_CCL
 from models.tt_transformers.tt.model_config import ModelArgs
 from models.utility_functions import comp_allclose, comp_pcc, skip_for_grayskull
+from models.experimental.gemma3_4b.tests.references import reference_vision_encoder
 
 from models.experimental.gemma3_4b.tt.gemma_image_transformer import TtGemmaImageTransformer
 
@@ -54,7 +55,7 @@ def test_image_transformer_inference(batch, num_chunks, mesh_device):
     dim = model_args.vision_dim
     seq_len = model_args.image_size
 
-    reference_model = model_args.reference_vision_encoder()
+    reference_model = reference_vision_encoder(model_args)
     # reference_model.load_state_dict(partial_state_dict)
     reference_model.eval()
