+++ models/experimental/gemma3_4b/tests/vision_tests/test_vision_transformer_block.py
@@ -15,6 +15,7 @@ from models.tt_transformers.tt.ccl import TT_CCL
 from models.tt_transformers.tt.model_config import ModelArgs
 from models.experimental.gemma3_4b.tt.gemma_image_block import TtGemmaImageTransformerBlock
 from models.utility_functions import comp_allclose, comp_pcc, skip_for_grayskull
+from models.experimental.gemma3_4b.tests.references import reference_vision_encoder_block
 
 
 @skip_for_grayskull("Requires wormhole_b0 to run")
@@ -57,7 +58,7 @@ def test_block_inference(batch, num_chunks, mesh_device, reset_seeds, gated):
     heads = model_args.vision_attn_n_heads
     seq_len = model_args.image_size
 
-    reference_model = model_args.reference_vision_encoder_block()
+    reference_model = reference_vision_encoder_block(model_args)
     # reference_model.load_state_dict(partial_state_dict)
     reference_model.eval()
